Created feature folders and starter files in prep for open sourcing the project
Each command for blkmrln has now been mapped out to a feature folder in src and a corresponding testing folder in test. When new developers start working on the project the expectation is that all work will be done in the feature folders and test folders until a pull request is created at which time it is evaluated if the PR will be merged.